Use serialization to save object state
Serialization is the process of converting an object into a stream of bytes to store the object or transmit it to memory, a database, or a file. Its main purpose is to save the state of an object in order to be able to recreate it when needed. The reverse process is called deserialization(unserialization in our case). Thanks docs.microsoft.com :smile: